move content areas to module
